ADDED: ModalAddUser
Also made the User.id optional ("?:") set the type of
new users to "User". The docRef.id  of the new user that
was successfully saved to firebase is then passed to
ModalUserDetails, which in turn shows the details of
the newly-created user.

RE Input Validation:
There's form validation happening under the hood, but
right now isn't shown to the user-end just yet.

Other notes:
Optimizations and bugfixes were made in ModalEditUser, UserList
Also removed USE_MOCK in UserList