Search by fullName
CURRENT LIMITATION:
You have to be exact in your search terms, i.e. if you're searching for "Perry
Ramos" you have to type in the whole phrase "Perry Ramos" and not just "perry" or "ramos". Otherwise, you won't get any result.

I now realize the importance of designing my user model to have a separate first name and last name for easier search purposes.  If only we could get an "or" for Firestore queries, our job here's done.  Currently, there is no support for that, see the ticket below:

https://github.com/firebase/firebase-js-sdk/issues/321

An alternate solution would be ElasticSearch, but since this feature isn't of a higher priority than finishing the whole module I'd settle with this for now.